Changed Zoom from int to double Implemented ZoomTypes
git-svn-id: http://devserv/svn/fluxdvd/components/wxPDFView/trunk@8137 ee4118f4-65c3-8c42-83c6-dbb17464b6b1

+++ include/PDFView.h
@@ -51,13 +51,13 @@ public:
 
 	int GetCurrentPage() const;
 
-	void SetZoom(int zoom);
+	void SetZoom(double zoom);
 
-	int GetZoom() const;
+	double GetZoom() const;
 
-	int GetMaxZoom() const;
+	double GetMaxZoom() const;
 
-	int GetMinZoom() const;
+	double GetMinZoom() const;
 
 	void SetZoomType(wxPDFViewZoomType zoomType);
